# openapi/renderweave-v1.yaml
openapi: 3.1.2
info:
  title: RenderWeave API
  version: 0.6.0
  description: >-
    Contract-first API for RenderWeave Schema Drafts, immutable StaticSchemas,
    exact compiled artifacts and authoritative RootDocument validation.
servers:
  - url: /
...
paths:
...
  /api/v1/schema-drafts:
    get:
      operationId: listDrafts
      tags: [Schema Drafts]
      summary: Page through active Schema Drafts, most recently updated first
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Active Draft summary page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftListResponse'
        '400':
          $ref: '#/components/responses/BadRequestProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
    post:
      operationId: createDraft
      tags: [Schema Drafts]
      summary: Create a valid Schema Draft at revision 0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDraftRequest'
      responses:
        '201':
          description: Draft created.
          headers:
            Location:
              description: Canonical URL of the created Draft.
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftResponse'
        '400':
          $ref: '#/components/responses/BadRequestProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '422':
          $ref: '#/components/responses/ValidationProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
...
  /api/v1/static-schemas:
    get:
      operationId: listStaticSchemas
      tags: [Static Schemas]
      summary: List immutable StaticSchemas, including system presets
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: StaticSchema summary page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaticSchemaListResponse'
        '400':
          $ref: '#/components/responses/BadRequestProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
    post:
      operationId: publishStaticSchema
      tags: [Static Schemas]
      summary: Publish the exact current Draft revision as an immutable StaticSchema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishStaticSchemaRequest'
      responses:
        '201':
          description: StaticSchema published with a stored compiled artifact.
          headers:
            Location:
              description: Canonical URL of the StaticSchema.
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaticSchemaResponse'
        '400':
          $ref: '#/components/responses/BadRequestProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '422':
          $ref: '#/components/responses/ValidationProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
  /api/v1/static-schemas/{schemaKey}/{versionTag}:
    parameters:
      - $ref: '#/components/parameters/StaticSchemaKey'
      - $ref: '#/components/parameters/VersionTag'
    get:
      operationId: getStaticSchema
      tags: [Static Schemas]
      summary: Read immutable StaticSchema metadata and DSL snapshot
      responses:
        '200':
          description: StaticSchema detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaticSchemaResponse'
        '400':
          $ref: '#/components/responses/BadRequestProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
  /api/v1/static-schemas/{schemaKey}/{versionTag}/definition:
    parameters:
      - $ref: '#/components/parameters/StaticSchemaKey'
      - $ref: '#/components/parameters/VersionTag'
    get:
      operationId: downloadStaticSchemaDefinition
      tags: [Static Schemas]
      summary: Download the immutable RenderWeave DSL snapshot
      responses:
        '200':
          description: Normalized RenderWeave DSL JSON.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaDefinition'
        '400':
          $ref: '#/components/responses/BadRequestProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
  /api/v1/static-schemas/{schemaKey}/{versionTag}/compiled-json-schema:
    parameters:
      - $ref: '#/components/parameters/StaticSchemaKey'
      - $ref: '#/components/parameters/VersionTag'
    get:
      operationId: downloadCompiledJsonSchema
      tags: [Static Schemas]
      summary: Download the exact stored JSON Schema 2020-12 artifact bytes
      responses:
        '200':
          description: Fully inlined immutable JSON Schema artifact.
          content:
            application/schema+json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequestProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
  /api/v1/static-schemas/{schemaKey}/{versionTag}/copies:
    parameters:
      - $ref: '#/components/parameters/StaticSchemaKey'
      - $ref: '#/components/parameters/VersionTag'
    post:
      operationId: copyStaticSchemaToDraft
      tags: [Static Schemas]
      summary: Copy one exact StaticSchema definition into a new user Draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CopyDraftRequest'
      responses:
        '201':
          description: New Draft at revision 0.
          headers:
            Location:
              description: Canonical URL of the new Draft.
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftResponse'
        '400':
          $ref: '#/components/responses/BadRequestProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '422':
          $ref: '#/components/responses/ValidationProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
...
components:
  parameters:
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    StaticSchemaKey:
      name: schemaKey
      in: path
      required: true
      description: User or reserved system StaticSchema key.
      schema:
        type: string
        pattern: '^[a-z0-9][a-z0-9-]{0,62}$'
    VersionTag:
      name: versionTag
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9][a-z0-9._-]{0,63}$'
  schemas:
...
    CreateDraftRequest:
      type: object
      additionalProperties: false
      required: [schemaKey, definition]
      properties:
        schemaKey:
          type: string
          pattern: '^[a-z0-9][a-z0-9-]{0,62}$'
        definition:
          $ref: '#/components/schemas/SchemaDefinition'
...
    CopyDraftRequest:
      type: object
      additionalProperties: false
      required: [schemaKey, displayName]
      properties:
        schemaKey:
          type: string
          pattern: '^[a-z0-9][a-z0-9-]{0,62}$'
        displayName:
          type: string
          minLength: 1
          maxLength: 128
    PublishStaticSchemaRequest:
      type: object
      additionalProperties: false
      required: [schemaKey, expectedRevision, versionTag]
      properties:
        schemaKey:
          type: string
          pattern: '^[a-z0-9][a-z0-9-]{0,62}$'
        expectedRevision:
          type: integer
          format: int64
          minimum: 0
        versionTag:
          type: string
          pattern: '^[a-z0-9][a-z0-9._-]{0,63}$'
        releaseNote:
          type: string
    DraftResponse:
      type: object
      additionalProperties: false
      required:
        - schemaKey
        - revision
        - definition
        - creationSource
        - createdAt
        - updatedAt
        - savedAt
        - resolvedRevisions
      properties:
        schemaKey:
          type: string
        revision:
          type: integer
          format: int64
          minimum: 0
        definition:
          $ref: '#/components/schemas/SchemaDefinition'
        creationSource:
          type: string
          enum: [USER, AI]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        savedAt:
          type: string
          format: date-time
        resolvedRevisions:
          type: object
          description: Root-inclusive request snapshot of live Draft revisions.
          additionalProperties:
            type: integer
            format: int64
            minimum: 0
    DraftListResponse:
      type: object
      additionalProperties: false
      required: [items, page, size, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DraftSummary'
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          format: int64
          minimum: 0
    DraftSummary:
      type: object
      additionalProperties: false
      required: [schemaKey, revision, creationSource, displayName, fieldCount, createdAt, updatedAt, savedAt]
      properties:
        schemaKey:
          type: string
        revision:
          type: integer
          format: int64
          minimum: 0
        creationSource:
          type: string
          enum: [USER, AI]
        displayName:
          type: string
        fieldCount:
          type: integer
          minimum: 0
          maximum: 256
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        savedAt:
          type: string
          format: date-time
...
    StaticSchemaResponse:
      type: object
      additionalProperties: false
      required:
        - schemaKey
        - versionTag
        - origin
        - sourceDraftRevision
        - definition
        - compilerVersion
        - releaseNote
        - referenceDepth
        - publishedAt
      properties:
        schemaKey:
          type: string
        versionTag:
          type: string
        origin:
          type: string
          enum: [DRAFT, SYSTEM]
        sourceDraftRevision:
          type: [integer, 'null']
          format: int64
          minimum: 0
        definition:
          $ref: '#/components/schemas/SchemaDefinition'
        compilerVersion:
          type: string
        releaseNote:
          type: [string, 'null']
        referenceDepth:
          type: integer
          minimum: 1
          maximum: 16
        publishedAt:
          type: string
          format: date-time
    StaticSchemaListResponse:
      type: object
      additionalProperties: false
      required: [items, page, size, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StaticSchemaSummaryResponse'
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          format: int64
          minimum: 0
    StaticSchemaSummaryResponse:
      type: object
      additionalProperties: false
      required:
        - schemaKey
        - versionTag
        - origin
        - displayName
        - fieldCount
        - referenceDepth
        - publishedAt
      properties:
        schemaKey: { type: string }
        versionTag: { type: string }
        origin:
          type: string
          enum: [DRAFT, SYSTEM]
        displayName: { type: string }
        fieldCount:
          type: integer
          minimum: 0
          maximum: 256
        referenceDepth:
          type: integer
          minimum: 1
          maximum: 16
        publishedAt:
          type: string
          format: date-time
    SchemaDefinition:
      type: object
      additionalProperties: false
      required: [dslVersion, displayName, fields]
      properties:
        dslVersion:
          type: string
          const: renderweave-schema/1.0
        displayName:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type: string
          maxLength: 2048
        fields:
          type: array
          maxItems: 256
          items:
            $ref: '#/components/schemas/SchemaField'
    SchemaField:
      type: object
      additionalProperties: false
      required: [fieldKey, required, value]
      properties:
        fieldKey:
          type: string
          minLength: 1
          x-renderweave-max-utf8-bytes: 128
        displayName:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type: string
          maxLength: 2048
        required:
          type: boolean
        value:
          $ref: '#/components/schemas/ValueDescriptor'
    ValueDescriptor:
      oneOf:
        - $ref: '#/components/schemas/TextValue'
        - $ref: '#/components/schemas/DecimalValue'
        - $ref: '#/components/schemas/DateValue'
        - $ref: '#/components/schemas/TimeValue'
        - $ref: '#/components/schemas/BooleanValue'
        - $ref: '#/components/schemas/ReferenceValue'
        - $ref: '#/components/schemas/ArrayValue'
    TextValue:
      type: object
      additionalProperties: false
      required: [type]
      properties:
        type:
          type: string
          const: text
        constraints:
          $ref: '#/components/schemas/TextConstraints'
    TextConstraints:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        minLength:
          type: integer
          minimum: 0
          maximum: 65536
        maxLength:
          type: integer
          minimum: 0
          maximum: 65536
        pattern:
          type: string
          maxLength: 1024
        enum:
          type: array
          minItems: 1
          maxItems: 256
          uniqueItems: true
          items:
            type: string
        const:
          type: string
    DecimalValue:
      type: object
      additionalProperties: false
      required: [type]
      properties:
        type:
          type: string
          const: decimal
        constraints:
          $ref: '#/components/schemas/DecimalConstraints'
    DecimalConstraints:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        min: { type: number }
        exclusiveMin: { type: number }
        max: { type: number }
        exclusiveMax: { type: number }
        multipleOf: { type: number, exclusiveMinimum: 0 }
        enum:
          type: array
          minItems: 1
          maxItems: 256
          uniqueItems: true
          items: { type: number }
        const: { type: number }
    DateValue:
      type: object
      additionalProperties: false
      required: [type]
      properties:
        type: { type: string, const: date }
        constraints:
          $ref: '#/components/schemas/DateConstraints'
    DateConstraints:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        min: { $ref: '#/components/schemas/DateLiteral' }
        exclusiveMin: { $ref: '#/components/schemas/DateLiteral' }
        max: { $ref: '#/components/schemas/DateLiteral' }
        exclusiveMax: { $ref: '#/components/schemas/DateLiteral' }
        enum:
          type: array
          minItems: 1
          maxItems: 256
          uniqueItems: true
          items: { $ref: '#/components/schemas/DateLiteral' }
        const: { $ref: '#/components/schemas/DateLiteral' }
    DateLiteral:
      type: string
      pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}$'
    TimeValue:
      type: object
      additionalProperties: false
      required: [type]
      properties:
        type: { type: string, const: time }
        constraints:
          $ref: '#/components/schemas/TimeConstraints'
    TimeConstraints:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        min: { $ref: '#/components/schemas/TimeLiteral' }
        exclusiveMin: { $ref: '#/components/schemas/TimeLiteral' }
        max: { $ref: '#/components/schemas/TimeLiteral' }
        exclusiveMax: { $ref: '#/components/schemas/TimeLiteral' }
        enum:
          type: array
          minItems: 1
          maxItems: 256
          uniqueItems: true
          items: { $ref: '#/components/schemas/TimeLiteral' }
        const: { $ref: '#/components/schemas/TimeLiteral' }
    TimeLiteral:
      type: string
      pattern: '^(?:[01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]$'
    BooleanValue:
      type: object
      additionalProperties: false
      required: [type]
      properties:
        type: { type: string, const: boolean }
        constraints:
          $ref: '#/components/schemas/BooleanConstraints'
    BooleanConstraints:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        const: { type: boolean }
    ReferenceValue:
      type: object
      additionalProperties: false
      required: [type, ref]
      properties:
        type: { type: string, const: reference }
        ref:
          $ref: '#/components/schemas/SchemaReference'
    SchemaReference:
      oneOf:
        - $ref: '#/components/schemas/SchemaRef'
        - $ref: '#/components/schemas/StaticSchemaRef'
    SchemaRef:
      type: object
      additionalProperties: false
      required: [schemaKey]
      properties:
        schemaKey:
          type: string
          pattern: '^[a-z0-9][a-z0-9-]{0,62}$'
    StaticSchemaRef:
      type: object
      additionalProperties: false
      required: [schemaKey, versionTag]
      properties:
        schemaKey:
          type: string
          pattern: '^[a-z0-9][a-z0-9-]{0,62}$'
        versionTag:
          type: string
          pattern: '^[a-z0-9][a-z0-9._-]{0,63}$'
    ArrayValue:
      type: object
      additionalProperties: false
      required: [type, items]
      properties:
        type: { type: string, const: array }
        constraints:
          $ref: '#/components/schemas/ArrayConstraints'
        items:
          oneOf:
            - $ref: '#/components/schemas/TextValue'
            - $ref: '#/components/schemas/DecimalValue'
            - $ref: '#/components/schemas/DateValue'
            - $ref: '#/components/schemas/TimeValue'
            - $ref: '#/components/schemas/BooleanValue'
            - $ref: '#/components/schemas/ReferenceValue'
    ArrayConstraints:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        minItems:
          type: integer
          minimum: 0
          maximum: 10000
        maxItems:
          type: integer
          minimum: 0
          maximum: 10000
        uniqueItems:
          type: boolean
...
    Violation:
      type: object
      additionalProperties: false
      required: [code, pointer, args]
      properties:
        code:
          type: string
        pointer:
          type: string
          description: RFC 6901 JSON Pointer relative to the invalid request object.
        args:
          type: object
          additionalProperties: true
        message:
          type: string
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
        traceId:
          type: string
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
        revision:
          type: integer
          format: int64
          minimum: 0
  responses:
    BadRequestProblem:
      description: Request JSON, key syntax or envelope is invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFoundProblem:
      description: The requested Draft does not exist.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ConflictProblem:
      description: Natural identity or expected revision conflicts with current state.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationProblem:
      description: The complete RenderWeave definition is not valid and nothing was written.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    InternalProblem:
      description: RFC 9457 problem response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'